Updated Const class
